move switchNetwork to web3-helpers
