+++ src/components/WalletAuthModal.tsx
@@ -3,7 +3,7 @@ import { useState, useEffect } from 'react';
 import { Web3Provider } from '@ethersproject/providers';
 import { useWeb3React } from '@web3-react/core';
 import { WalletConnectConnector } from '@web3-react/walletconnect-connector';
-import { loginSupportedChainIds } from 'lib/login';
+import { loginSupportedChainIds, switchNetwork } from 'lib/web3-helpers';
 import { concat } from 'lodash';
 
 import {
