+++ src/hooks/useConnectedChain.ts
@@ -1,5 +1,5 @@
 import { useWeb3React } from '@web3-react/core';
-import { loginSupportedChainIds } from 'lib/login';
+import { loginSupportedChainIds } from 'lib/web3-helpers';
 
 export default function useConnectedChain() {
   const { chainId } = useWeb3React();
