--- src/lib/login.ts
@@ -1,10 +1,0 @@
-export const loginSupportedChainIds: { [key: number]: string } = {
-  1: 'Ethereum Mainnet',
-  10: 'Optimism',
-  100: 'Gnosis',
-  137: 'Polygon',
-  250: 'Fantom Opera',
-  42220: 'Celo Mainnet',
-  42161: 'Arbitrum One',
-  43114: 'Avalanche C-Chain',
-};
